Fix cookie coloring.
